Added google tag to be able to see search terms
Added a file to verify ownership of tribuo.org to google.  This will allow us to receive search terms that are used in google searches that lead to our site. No data about the user is stored by us, just that the search terms lead to tribuo.org. File should remain in place until we have a different option for analytics.